deployment/docker/ml/requirements.txt
colorama==0.4.4
joblib
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
gunicorn==20.1.0
requests==2.27.1
flask_wtf==1.0.0
peewee==3.14.10
wtforms
pymysql
click==8.1.3
jupyterlab==3.4.7
matplotlib==3.5.3
datasets==2.5.1
pandas==1.4.4
plotly==5.10.0
pyspark==3.3.0
pytorch-lightning==1.7.7
scapy[complete]==2.5.0rc1
scikit-learn==1.1.2
seaborn==0.11.2
tensorboard==2.10.0
--extra-index-url https://download.pytorch.org/whl/cpu
--pre
- torch==1.12.1
- torchvision==0.13.1
- torchaudio==0.12.1
torchmetrics==0.11.4
